+++ functions/index.js
@@ -1,2 +1,2 @@
 exports.inbound = require('./inbound');
-exports.airtableTriager = require('./airtableTriager');
+exports.poll = require('./poll');
