+++ functions/poll.js
@@ -1,0 +1,85 @@
+const functions = require('firebase-functions');
+const Slack = require('slack');
+
+const { getChangedIntakeTickets } = require('./airtable');
+
+const bot = new Slack({ token: functions.config().slack.token });
+
+const NEIGHBORHOOD_TO_CHANNEL = {
+  NW: 'northwest_bedstuy',
+  NE: 'northeast_bedstuy',
+  SW: 'southwest_bedstuy',
+  SE: 'southeast_bedstuy',
+  // "Clinton Hill / Fort Greene": "clintonhill",
+  // "Crown Heights / Brownsville / Flatbush": "crownheights",
+};
+
+const CHANNEL_TO_ID = functions.config().slack.channel_to_id;
+
+async function onNewIntake(id, fields) {
+  console.log(`onNewIntake(${id})`);
+
+  const neighborhoodChannelName = NEIGHBORHOOD_TO_CHANNEL[fields['Neighborhood']];
+
+  // e.g. crown heights
+  if (!neighborhoodChannelName) {
+    return;
+  }
+
+  const neighborhoodChannelID = CHANNEL_TO_ID[neighborhoodChannelName];
+
+  // TODO : add some better content
+  const res = await bot.chat.postMessage({
+    channel: neighborhoodChannelID,
+    text: `HELLO! There is a new ticket: ${fields['Ticket ID']}`,
+  });
+
+  if (!res['ok']) {
+    console.error(`Encountered an error posting ticket to #${neighborhoodChannelName}: ${id}`);
+    return;
+  }
+}
+
+module.exports = {
+  // Runs every minute
+  airtableIntakeTickets: functions.pubsub.schedule('* * * * *').onRun(async () => {
+    const STATUS_TO_CBS = {
+      'Seeking Volunteer': [onNewIntake],
+      'Assigned / In Progress': [],
+      'Complete': [],
+      'Not Bed-Stuy': [],
+    };
+
+    const changedTickets = await getChangedIntakeTickets();
+
+    if (changedTickets.length === 0) {
+      return null;
+    }
+
+    // TODO : it is possible for us to miss a step in the intake ticket state transitions.
+    // A ticket should go from "Seeking Volunteer" -> "Assigned" -> "Complete". Since we
+    // only trigger on the current state, there is a race condition where we could miss
+    // the intermediate state (i.e. assigned)
+    const callbacks = [];
+    for (const [id, fields] of changedTickets) {
+      console.log(`Processing intake ticket (${fields['Ticket ID']}): ${id}`);
+
+      const status = fields['Status'];
+      if (!(status in STATUS_TO_CBS)) {
+        console.error(`Ticket has invalid status: ${status}`);
+        continue;
+      }
+
+      for (const callback of STATUS_TO_CBS[status]) {
+        // eslint-disable-next-line callback-return
+        callbacks.push(callback(id, fields));
+      }
+    }
+    await Promise.all(callbacks);
+
+    // TODO : disable this logging later, might be useful as we get this up and running
+    console.log(`All changed tickets: ${changedTickets}`);
+
+    return null;
+  }),
+}
